Renamed some files and fixed the GUI to interact with the GraphManager

+++ SQLSkillBuilder.cs
@@ -11,7 +11,7 @@ namespace SkillSimulator
 
         public Skill BuildSkill(object[] rawdata)
         {
-            int id = (int)rawdata[0];
+            int id = (short)rawdata[0];
             String name = (String)rawdata[1];
             short maxlvl = (short)rawdata[2];
             Skill skill = new Skill(id, name, maxlvl);
